Support for complex values, bug fixes
Replaced instances of the complex conjugate transpose (') with the scalar transpose (.') in order to properly support complex values. No test has be done in this area.

Also, zero-crossing event detection is now strict: tangents are not detected. This is to avoid multiple detects and unwanted detects at the initial conditions. This could be an option or somehow specified via the EventsFUN in the future.

Signed-off-by: Andrew D. Horchler <adh9@case.edu>