Refactored Ornstein-Uhlenbeck process function sde_ou(), zero drift case handled
Re wrote most of sde_ou(). Hopefully a bit cleaner and simpler. Still messy. Thanks to Kendrick (kms15) for the inspiration to reorganize some of the terms in the analytical solution - this simplifies some of the vector cases.

Fixed small bug in sde_gbm().

Fixed bug in sdearguments() that caused integration functions to use identical noise across dimensions in some cases.

Signed-off-by: Andrew D. Horchler <adh9@case.edu>